Add support for prism
