+++ src/AtomUI.Controls.DataGrid/ThemeManagerBuilderExtensions.cs
@@ -7,21 +7,18 @@ public static class OSSDataGridThemeManagerBuilderExtensions
 {
     public static IThemeManagerBuilder UseOSSDataGrid(this IThemeManagerBuilder themeManagerBuilder)
     {
-        themeManagerBuilder.AppBuilder.AfterSetup(_ =>
+        var controlTokenTypes = ControlTokenTypePool.GetTokenTypes();
+        foreach (var controlType in controlTokenTypes)
         {
-            var controlTokenTypes = ControlTokenTypePool.GetTokenTypes();
-            foreach (var controlType in controlTokenTypes)
-            {
-                themeManagerBuilder.AddControlToken(controlType);
-            }
-            themeManagerBuilder.AddControlThemesProvider(new AtomUIDataGridThemesProvider());
+            themeManagerBuilder.AddControlToken(controlType);
+        }
+        themeManagerBuilder.AddControlThemesProvider(new AtomUIDataGridThemesProvider());
 
-            var languageProviders = LanguageProviderPool.GetLanguageProviders();
-            foreach (var languageProvider in languageProviders)
-            {
-                themeManagerBuilder.AddLanguageProviders(languageProvider);
-            }
-        });
+        var languageProviders = LanguageProviderPool.GetLanguageProviders();
+        foreach (var languageProvider in languageProviders)
+        {
+            themeManagerBuilder.AddLanguageProviders(languageProvider);
+        }
         return themeManagerBuilder;
     }
 }
